Replace signature permission with whitelist
Instead of using a signature permission to protect
the service, we now use a whitelist approach to
ease the integration of the extension into custom roms.
Now rom maintainer can build the extension on their
own and sign it with their own key.
The extension then still works with a F-Droid client
signed with the F-Droid key.