also build update.zip that includes stable release FDroid.apk
Include FDroid.apk in /system/app/FDroid/, where it will not be
granted any particular permissions, but will be shipped as part of the
system.  This is useful for OEMs who are shipped complete devices and
want to just flash a single ZIP in order to install F-Droid by
default.

https://android.googlesource.com/platform/frameworks/base/+/ccbf84f

closes #7