use generated versionName for gradle builds to ease IDing builds
If this project is successful, lots of people will be building it and
flashing it themselves, as part of ROMs or other experiements.  It is very
likely that they will not change the versionName, so here we can generate
it using `git describe` so that it is a lot easier to figure out where the
code came from.

The versionName is still hardcoded in AndroidManifest.xml until we can
figure out how to set it using `git describe` in an Android.mk.

#8