use zopflipng for smaller PNGs and reproducible builds
PNG crunching is not a deterministic process, especially the way aapt does
it.  This makes the F-Droid builds not reproducible.  The easy solution to
this is to pre-crunch the PNGs and commit them to git.  This also then lets
us use zipflipng, which makes for a lot smaller PNGs than the built-in
method.

https://medium.com/@duhroach/smaller-pngs-and-android-s-aapt-tool-4ce38a24019d
https://ariya.io/2016/06/using-zopfli-to-optimize-png-images

I built zopflipng on Ubuntu/xenial like this:

apt install zopfli
git clone https://github.com/google/zopfli
cd zopfli
make zopflipng